make migrations rely on flyway instead of hibernate auto-creation

+++ api/src/main/java/com/exelaration/abstractmemery/services/MetadataService.java
@@ -8,9 +8,9 @@ public interface MetadataService {
 
   public List<Image> findAll();
 
-  public Optional<Image> findById(Long id);
+  public Optional<Image> findById(Integer id);
 
   public Image save(Image image);
 
-  public void deleteById(Long id);
+  public void deleteById(Integer id);
 }
